+++ kapacitor/client.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"fmt"
 	"net/http"
-	"regexp"
-	"strings"
 
 	"github.com/influxdata/chronograf"
 	"github.com/influxdata/chronograf/id"
@@ -68,10 +66,13 @@ type Task struct {
 	TICKScript chronograf.TICKScript // TICKScript is the running script
 }
 
-var reTaskName = regexp.MustCompile(`[\r\n]*var[ \t]+name[ \t]+=[ \t]+'([^\n]+)'[ \r\t]*\n`)
-
 // NewTask creates a task from a kapacitor client task
 func NewTask(task *client.Task) *Task {
+	return NewTaskWithParsing(task, true)
+}
+
+// NewTask creates a task from a kapacitor client task with optional parsing of alert rule
+func NewTaskWithParsing(task *client.Task, parse bool) *Task {
 	dbrps := make([]chronograf.DBRP, len(task.DBRPs))
 	for i := range task.DBRPs {
 		dbrps[i].DB = task.DBRPs[i].Database
@@ -80,26 +81,20 @@ func NewTask(task *client.Task) *Task {
 
 	script := chronograf.TICKScript(task.TICKscript)
 	var rule chronograf.AlertRule = chronograf.AlertRule{Query: nil}
-	if task.TemplateID == "" {
+	if parse && task.TemplateID == "" {
 		// try to parse chronograf rule, tasks created from template cannot be chronograf rules
 		if parsedRule, err := Reverse(script); err == nil {
 			rule = parsedRule
+			// #5403 override name when defined in a variable
+			if nameVar, exists := task.Vars["name"]; exists {
+				if val, isString := nameVar.Value.(string); isString && val != "" {
+					rule.Name = val
+				}
+			}
 		}
 	}
 	if rule.Name == "" {
-		// try to parse Name from a line such as: `var name = 'Rule Name'
-		if matches := reTaskName.FindStringSubmatch(task.TICKscript); matches != nil {
-			rule.Name = strings.ReplaceAll(strings.ReplaceAll(matches[1], "\\'", "'"), "\\\\", "\\")
-		} else {
-			rule.Name = task.ID
-		}
-	}
-
-	// #5403 override name when defined in a variable
-	if nameVar, exists := task.Vars["name"]; exists {
-		if val, isString := nameVar.Value.(string); isString && val != "" {
-			rule.Name = val
-		}
+		rule.Name = GetAlertRuleName(task)
 	}
 
 	rule.Vars = task.Vars
@@ -279,13 +274,16 @@ func (c *Client) status(ctx context.Context, href string) (client.TaskStatus, er
 
 // All returns all tasks in kapacitor
 func (c *Client) All(ctx context.Context) (map[string]*Task, error) {
+	return c.List(ctx, &client.ListTasksOptions{}, true)
+}
+
+// List kapacitor tasks according to options supplied
+func (c *Client) List(ctx context.Context, opts *client.ListTasksOptions, parseRules bool) (map[string]*Task, error) {
 	kapa, err := c.kapaClient(c.URL, c.Username, c.Password, c.InsecureSkipVerify)
 	if err != nil {
 		return nil, err
 	}
 
-	// Only get the status, id and link section back
-	opts := &client.ListTasksOptions{}
 	tasks, err := kapa.ListTasks(opts)
 	if err != nil {
 		return nil, err
@@ -293,7 +291,7 @@ func (c *Client) All(ctx context.Context) (map[string]*Task, error) {
 
 	all := map[string]*Task{}
 	for _, task := range tasks {
-		all[task.ID] = NewTask(&task)
+		all[task.ID] = NewTaskWithParsing(&task, parseRules)
 	}
 	return all, nil
 }
@@ -325,7 +323,6 @@ func (c *Client) Get(ctx context.Context, id string) (*Task, error) {
 	if err != nil {
 		return nil, chronograf.ErrAlertNotFound
 	}
-	fmt.Println("!!!", task.ID, task.TemplateID)
 
 	return NewTask(&task), nil
 }