+++ kapacitor/kapa_client.go
@@ -1,6 +1,9 @@
 package kapacitor
 
 import (
+	"errors"
+	"regexp"
+	"strings"
 	"sync"
 
 	client "github.com/influxdata/kapacitor/client/v1"
@@ -11,12 +14,9 @@ const (
 	// tasks from Kapacitor. This constant was chosen after some benchmarking
 	// work and should likely work well for quad-core systems
 	ListTaskWorkers = 4
-
-	// TaskGatherers is the number of workers collating responses from
-	// ListTaskWorkers. There can only be one without additional synchronization
-	// around the output buffer from ListTasks
-	TaskGatherers = 1
 )
+
+const maxLimit = 1_000_000
 
 // ensure PaginatingKapaClient is a KapaClient
 var _ KapaClient = &PaginatingKapaClient{}
@@ -31,18 +31,21 @@ type PaginatingKapaClient struct {
 // ListTasks lists all available tasks from Kapacitor, navigating pagination as
 // it fetches them
 func (p *PaginatingKapaClient) ListTasks(opts *client.ListTasksOptions) ([]client.Task, error) {
-	// only trigger auto-pagination with Offset=0 and Limit=0
-	if opts.Limit != 0 || opts.Offset != 0 {
-		return p.KapaClient.ListTasks(opts)
+	if opts.Pattern != "" && opts.Offset > 0 {
+		return nil, errors.New("offset paramater must be 0 when pattern parameter is supplied")
 	}
-
-	allTasks := []client.Task{}
+	allTasks := make([]client.Task, 0, p.FetchRate)
 
 	optChan := make(chan client.ListTasksOptions)
 	taskChan := make(chan []client.Task, ListTaskWorkers)
 	done := make(chan struct{})
 
 	var once sync.Once
+	readFinished := func() {
+		once.Do(func() {
+			close(done)
+		})
+	}
 
 	go p.generateKapacitorOptions(optChan, *opts, done)
 
@@ -50,14 +53,41 @@ func (p *PaginatingKapaClient) ListTasks(opts *client.ListTasksOptions) ([]clien
 
 	wg.Add(ListTaskWorkers)
 	for i := 0; i < ListTaskWorkers; i++ {
-		go p.fetchFromKapacitor(optChan, &wg, &once, taskChan, done)
+		go func() {
+			p.fetchFromKapacitor(optChan, taskChan, readFinished)
+			wg.Done()
+		}()
 	}
 
 	var gatherWg sync.WaitGroup
-	gatherWg.Add(TaskGatherers)
+	gatherWg.Add(1)
 	go func() {
-		for task := range taskChan {
-			allTasks = append(allTasks, task...)
+		// pos and limit are position within filtered results
+		// that are used with pattern searching
+		pos := -1
+		limit := opts.Limit
+		if limit <= 0 {
+			limit = maxLimit
+		}
+	processTasksBatches:
+		for tasks := range taskChan {
+			if opts.Pattern != "" {
+				for _, task := range tasks {
+					if strings.Contains(GetAlertRuleName(&task), opts.Pattern) {
+						pos++
+						if pos < opts.Offset {
+							continue
+						}
+						allTasks = append(allTasks, task)
+						if len(allTasks) == limit {
+							readFinished()
+							break processTasksBatches
+						}
+					}
+				}
+				continue
+			}
+			allTasks = append(allTasks, tasks...)
 		}
 		gatherWg.Done()
 	}()
@@ -72,8 +102,7 @@ func (p *PaginatingKapaClient) ListTasks(opts *client.ListTasksOptions) ([]clien
 // fetchFromKapacitor fetches a set of results from a kapacitor by reading a
 // set of options from the provided optChan. Fetched tasks are pushed onto the
 // provided taskChan
-func (p *PaginatingKapaClient) fetchFromKapacitor(optChan chan client.ListTasksOptions, wg *sync.WaitGroup, closer *sync.Once, taskChan chan []client.Task, done chan struct{}) {
-	defer wg.Done()
+func (p *PaginatingKapaClient) fetchFromKapacitor(optChan chan client.ListTasksOptions, taskChan chan []client.Task, readFinished func()) {
 	for opt := range optChan {
 		resp, err := p.KapaClient.ListTasks(&opt)
 		if err != nil {
@@ -82,9 +111,7 @@ func (p *PaginatingKapaClient) fetchFromKapacitor(optChan chan client.ListTasksO
 
 		// break and stop all workers if we're done
 		if len(resp) == 0 {
-			closer.Do(func() {
-				close(done)
-			})
+			readFinished()
 			return
 		}
 
@@ -96,18 +123,64 @@ func (p *PaginatingKapaClient) fetchFromKapacitor(optChan chan client.ListTasksO
 // generateKapacitorOptions creates ListTasksOptions with incrementally greater
 // Limit and Offset parameters, and inserts them into the provided optChan
 func (p *PaginatingKapaClient) generateKapacitorOptions(optChan chan client.ListTasksOptions, opts client.ListTasksOptions, done chan struct{}) {
-	// ensure Limit and Offset start from known quantities
-	opts.Limit = p.FetchRate
-	opts.Offset = 0
+	toFetchCount := opts.Limit
+	if toFetchCount <= 0 {
+		toFetchCount = maxLimit
+	}
+	// fetch all data when pattern is set, chronograf is herein filters by task name
+	// whereas kapacitor filters by task ID that is hidden to chronograf users
+	if opts.Pattern != "" {
+		toFetchCount = maxLimit
+		opts.Pattern = ""
+		opts.Offset = 0
+	}
 
+	nextLimit := func() int {
+		if p.FetchRate < toFetchCount {
+			toFetchCount -= p.FetchRate
+			return p.FetchRate
+		}
+		retVal := toFetchCount
+		toFetchCount = 0
+		return retVal
+	}
+
+	opts.Limit = nextLimit()
+	opts.Pattern = ""
+
+generateOpts:
 	for {
 		select {
 		case <-done:
-			close(optChan)
-			return
+			break generateOpts
 		case optChan <- opts:
 			// nop
 		}
-		opts.Offset = p.FetchRate + opts.Offset
+		if toFetchCount <= 0 {
+			// no more data to read from options
+			break generateOpts
+		}
+		opts.Offset = opts.Offset + p.FetchRate
+		opts.Limit = nextLimit()
 	}
+	close(optChan)
+}
+
+var reTaskName = regexp.MustCompile(`[\r\n]*var[ \t]+name[ \t]+=[ \t]+'([^\n]+)'[ \r\t]*\n`)
+
+// GetAlertRuleName returns chronograf rule name out of kapacitor task.
+// Chronograf UI always preffer alert rule names.
+func GetAlertRuleName(task *client.Task) string {
+	// #5403 override name when defined in a variable
+	if nameVar, exists := task.Vars["name"]; exists {
+		if val, isString := nameVar.Value.(string); isString && val != "" {
+			return val
+		}
+	}
+	// try to parse Name from a line such as: `var name = 'Rule Name'
+	if matches := reTaskName.FindStringSubmatch(task.TICKscript); matches != nil {
+		return strings.ReplaceAll(strings.ReplaceAll(matches[1], "\\'", "'"), "\\\\", "\\")
+	}
+
+	return task.ID
 }