+++ ui/src/kapacitor/index.js
@@ -3,6 +3,7 @@ import KapacitorRulePage from './containers/KapacitorRulePage'
 import KapacitorRulesPage from './containers/KapacitorRulesPage'
 import TickscriptPage from './containers/TickscriptPage'
 import FluxTaskPage from './containers/FluxTaskPage'
+import FluxTasksPage from './containers/FluxTasksPage'
 
 export {
   KapacitorPage,
@@ -10,4 +11,5 @@ export {
   KapacitorRulesPage,
   TickscriptPage,
   FluxTaskPage,
+  FluxTasksPage,
 }