+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 1. [#5852](https://github.com/influxdata/chronograf/pull/5852): Add Flux Query Builder.
 1. [#5858](https://github.com/influxdata/chronograf/pull/5858): Use time range in flux Schema Explorer.
 1. [#5861](https://github.com/influxdata/chronograf/pull/5861): Allow to load truncated data in Flux Query Builder.
+1. [#5868](https://github.com/influxdata/chronograf/pull/5868): Move Flux Tasks to own page.
+1. [#5869](https://github.com/influxdata/chronograf/pull/5869): Optimize Alert Rules API.
 
 ### Bug Fixes
 
@@ -14,6 +16,7 @@
 
 1. [#5851](https://github.com/influxdata/chronograf/pull/5851): Remove fixed-table-2 to resolve CVE-2022-0235.
 1. [#5866](https://github.com/influxdata/chronograf/pull/5866): Replace axios by fetch.
+1. [#5873](https://github.com/influxdata/chronograf/pull/5873): Upgrade react-resize-detector.
 
 ## v1.9.3 [2022-01-25]